fix: clean deprecated methods
