+++ pkg/api/test_data/test_get_entries/with_path_entries/kcl.mod
@@ -5,4 +5,3 @@ version = "0.0.1"
 
 [profile]
 entries = ["sub/sub.k"]
-
