+++ pkg/api/test_data/test_kpm_package/get_schema_ty/aaa/kcl.mod.lock
@@ -3,4 +3,3 @@
     name = "bbb"
     full_name = "bbb_0.0.1"
     version = "0.0.1"
-    sum = "/PGJJE6Rp0lAsq3n9RUZ34jPGfwzYIZESs17L6kSG/Y="
