+++ pkg/api/test_data/test_kpm_package/kcl_pkg/kcl.mod.lock
@@ -3,3 +3,7 @@
     name = "helloworld"
     full_name = "helloworld_0.1.3"
     version = "0.1.3"
+    sum = "T9FGqbdtkuML4+pDyf4SFP5bDcDusbra6AKtCZygvNw="
+    reg = "ghcr.io"
+    repo = "kcl-lang/helloworld"
+    oci_tag = "0.1.3"
