+++ pkg/api/test_data/test_kpm_package/no_kcl_files/kcl.mod
@@ -2,4 +2,3 @@
 name = "no_kcl_files"
 edition = "0.0.1"
 version = "0.0.1"
-
