+++ pkg/api/test_data/test_work_dir/dev/kcl.mod
@@ -5,4 +5,3 @@ version = "0.1.0"
 
 [profile]
 entries = ["../base/base.k", "main.k"]
-
