+++ pkg/client/init.go
@@ -1,0 +1,53 @@
+package client
+
+import (
+	"fmt"
+	"path/filepath"
+
+	"kcl-lang.io/kpm/pkg/constants"
+	pkg "kcl-lang.io/kpm/pkg/package"
+	"kcl-lang.io/kpm/pkg/reporter"
+	"kcl-lang.io/kpm/pkg/utils"
+)
+
+// createIfNotExist will create a file if it does not exist.
+func (c *KpmClient) createIfNotExist(filepath string, storeFunc func() error) error {
+	reporter.ReportMsgTo(fmt.Sprintf("creating new :%s", filepath), c.GetLogWriter())
+	err := utils.CreateFileIfNotExist(
+		filepath,
+		storeFunc,
+	)
+	if err != nil {
+		if errEvent, ok := err.(*reporter.KpmEvent); ok {
+			if errEvent.Type() != reporter.FileExists {
+				return err
+			} else {
+				reporter.ReportMsgTo(fmt.Sprintf("'%s' already exists", filepath), c.GetLogWriter())
+			}
+		} else {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// InitEmptyPkg will initialize an empty kcl package.
+func (c *KpmClient) InitEmptyPkg(kclPkg *pkg.KclPkg) error {
+	err := c.createIfNotExist(kclPkg.ModFile.GetModFilePath(), kclPkg.ModFile.StoreModFile)
+	if err != nil {
+		return err
+	}
+
+	err = c.createIfNotExist(kclPkg.ModFile.GetModLockFilePath(), kclPkg.LockDepsVersion)
+	if err != nil {
+		return err
+	}
+
+	err = c.createIfNotExist(filepath.Join(kclPkg.ModFile.HomePath, constants.DEFAULT_KCL_FILE_NAME), kclPkg.CreateDefaultMain)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
