+++ pkg/client/push.go
@@ -1,0 +1,48 @@
+package client
+
+import (
+	"fmt"
+
+	"kcl-lang.io/kpm/pkg/oci"
+	"kcl-lang.io/kpm/pkg/opt"
+	"kcl-lang.io/kpm/pkg/reporter"
+	"kcl-lang.io/kpm/pkg/utils"
+)
+
+// PushToOci will push a kcl package to oci registry.
+func (c *KpmClient) PushToOci(localPath string, ociOpts *opt.OciOptions) error {
+	repoPath := utils.JoinPath(ociOpts.Reg, ociOpts.Repo)
+	cred, err := c.GetCredentials(ociOpts.Reg)
+	if err != nil {
+		return err
+	}
+
+	ociCli, err := oci.NewOciClientWithOpts(
+		oci.WithCredential(cred),
+		oci.WithRepoPath(repoPath),
+		oci.WithSettings(c.GetSettings()),
+		oci.WithInsecureSkipTLSverify(c.insecureSkipTLSverify),
+	)
+
+	if err != nil {
+		return err
+	}
+
+	ociCli.SetLogWriter(c.logWriter)
+
+	exist, err := ociCli.ContainsTag(ociOpts.Tag)
+	if err != (*reporter.KpmEvent)(nil) {
+		return err
+	}
+
+	if exist {
+		return reporter.NewErrorEvent(
+			reporter.PkgTagExists,
+			fmt.Errorf("package version '%s' already exists", ociOpts.Tag),
+		)
+	}
+
+	return ociCli.PushWithOciManifest(localPath, ociOpts.Tag, &opt.OciManifestOptions{
+		Annotations: ociOpts.Annotations,
+	})
+}
