+++ pkg/client/test_data/add_with_mod_spec/git/kcl.mod
@@ -1,0 +1,7 @@
+[package]
+name = "git"
+edition = "v0.10.0"
+version = "0.0.1"
+
+[dependencies]
+cc = { git = "https://github.com/kcl-lang/flask-demo-kcl-manifests.git", commit = "8308200", version = "0.0.1" }
