+++ pkg/client/test_data/add_with_mod_spec/git_mod_0/kcl.mod.lock
@@ -1,0 +1,7 @@
+[dependencies]
+  [dependencies.cc]
+    name = "cc"
+    full_name = "cc_0.0.1"
+    version = "0.0.1"
+    url = "https://github.com/kcl-lang/flask-demo-kcl-manifests.git"
+    commit = "8308200"
