+++ pkg/client/test_data/add_with_mod_spec/git_mod_1/kcl.mod
@@ -1,0 +1,4 @@
+[package]
+name = "git"
+edition = "v0.10.0"
+version = "0.0.1"
