+++ pkg/client/test_data/add_with_mod_spec/oci/kcl.mod
@@ -1,0 +1,7 @@
+[package]
+name = "oci"
+edition = "v0.10.0"
+version = "0.0.1"
+
+[dependencies]
+subhelloworld = { oci = "oci://ghcr.io/kcl-lang/helloworld", tag = "0.1.4", version = "0.0.1" }
