+++ pkg/client/test_data/add_with_mod_spec/oci/kcl.mod.lock
@@ -1,0 +1,9 @@
+[dependencies]
+  [dependencies.subhelloworld]
+    name = "subhelloworld"
+    full_name = "subhelloworld_0.0.1"
+    version = "0.0.1"
+    sum = "9J9HOMhdypaDYf0J7PqtpGTdlkbxkN0HFEYhosHhf4U="
+    reg = "ghcr.io"
+    repo = "kcl-lang/helloworld"
+    oci_tag = "0.1.4"
